requirements.txt
python-telegram-bot==20.7
notion-client==2.1.0
python-dotenv==1.0.0
